refactor: restructure text search APIs
1. Created new base classes TextSearchOptionsAPI and TextSearchResultAPI
2. Moved common functionality from ContentSearch and OcrTextSearch to
base classes
3. Changed ContentOptionsAPI and OcrTextOptionsAPI to inherit from
TextSearchOptionsAPI
4. Changed ContentResultAPI and OcrTextResultAPI to inherit from
TextSearchResultAPI
5. Consolidated duplicate code for preview length, highlighting, file
attributes and timestamps
6. Updated implementation files to reflect inheritance changes

The changes introduce better code organization by:
1. Removing duplicate code across different search types
2. Creating a clear hierarchy for search API classes
3. Making it easier to maintain and extend common functionality
4. Following DRY (Don't Repeat Yourself) principle
5. Increasing default preview length from 50 to 200 characters

All existing functionality remains the same, only the implementation
structure changed.

refactor: 重构文本搜索API结构

1. 创建新的基类TextSearchOptionsAPI和TextSearchResultAPI
2. 将ContentSearch和OcrTextSearch中的通用功能移至基类
3. 修改ContentOptionsAPI和OcrTextOptionsAPI继承自TextSearchOptionsAPI
4. 修改ContentResultAPI和OcrTextResultAPI继承自TextSearchResultAPI
5. 整合预览长度、高亮显示、文件属性和时间戳处理的重复代码
6. 更新实现文件以反映继承关系变化

这些改动通过以下方式改善了代码组织：
1. 消除了不同类型搜索间的重复代码
2. 为搜索API类创建了清晰的层次结构
3. 使得维护和扩展通用功能更加容易
4. 遵循DRY(不要重复自己)原则
5. 将默认预览长度从50字符增加到200字符

所有现有功能保持不变，仅改变了实现结构。
